+++ docs/apiref/classes/nullscrollbar.html
@@ -178,7 +178,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides E.__constructor</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/NullScrollbar.ts#L4">NullScrollbar.ts:4</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/NullScrollbar.ts#L4">NullScrollbar.ts:4</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -379,7 +379,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/scrollbaroperations.html">ScrollbarOperations</a>.<a href="../interfaces/scrollbaroperations.html#onchangebarpositionrate">onChangeBarPositionRate</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/NullScrollbar.ts#L4">NullScrollbar.ts:4</a></li>
+							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/NullScrollbar.ts#L4">NullScrollbar.ts:4</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -855,7 +855,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides E.destroy</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/NullScrollbar.ts#L11">NullScrollbar.ts:11</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/NullScrollbar.ts#L11">NullScrollbar.ts:11</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">void</span></h4>
@@ -1464,7 +1464,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/scrollbaroperations.html">ScrollbarOperations</a>.<a href="../interfaces/scrollbaroperations.html#setbarproperties">setBarProperties</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/NullScrollbar.ts#L16">NullScrollbar.ts:16</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/NullScrollbar.ts#L16">NullScrollbar.ts:16</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -1579,10 +1579,10 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="defaulthorizontalscrollbar.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar</a>
+						<a href="ninepatchhorizontalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="defaultverticalscrollbar.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar</a>
+						<a href="ninepatchverticalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar</a>
 					</li>
 				</ul>
 				<ul class="current">
@@ -1768,10 +1768,10 @@
 						<a href="scrolledcontentcontainer.html" class="tsd-kind-icon">Scrolled<wbr>Content<wbr>Container</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="../interfaces/defaulthorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="../interfaces/ninepatchhorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="../interfaces/defaultverticalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="../interfaces/ninepatchverticalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
 						<a href="../interfaces/scrollableparameterobject.html" class="tsd-kind-icon">Scrollable<wbr>Parameter<wbr>Object</a>
@@ -1780,7 +1780,7 @@
 						<a href="../interfaces/scrollbaroperations.html" class="tsd-kind-icon">Scrollbar<wbr>Operations</a>
 					</li>
 					<li class=" tsd-kind-type-alias tsd-is-external">
-						<a href="../globals.html#scrollbarlike" class="tsd-kind-icon">Scrollbar<wbr>Like</a>
+						<a href="../globals.html#scrollbar" class="tsd-kind-icon">Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-function tsd-is-external">
 						<a href="../globals.html#createdefaultscrollbarimage" class="tsd-kind-icon">create<wbr>Default<wbr>Scrollbar<wbr>Image</a>
