+++ docs/apiref/classes/scrolledcontent.html
@@ -176,7 +176,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides E.__constructor</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L8">Scrollable.ts:8</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L8">Scrollable.ts:8</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -376,7 +376,7 @@
 					<div class="tsd-signature tsd-kind-icon">on<wbr>Modified<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">Trigger</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">void</span><span class="tsd-signature-symbol">&gt;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L8">Scrollable.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L8">Scrollable.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -852,7 +852,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides E.destroy</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L13">Scrollable.ts:13</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L13">Scrollable.ts:13</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">void</span></h4>
@@ -1061,7 +1061,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides E.modified</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L17">Scrollable.ts:17</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L17">Scrollable.ts:17</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -1532,10 +1532,10 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="defaulthorizontalscrollbar.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar</a>
+						<a href="ninepatchhorizontalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="defaultverticalscrollbar.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar</a>
+						<a href="ninepatchverticalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
 						<a href="nullscrollbar.html" class="tsd-kind-icon">Null<wbr>Scrollbar</a>
@@ -1718,10 +1718,10 @@
 						<a href="scrolledcontentcontainer.html" class="tsd-kind-icon">Scrolled<wbr>Content<wbr>Container</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="../interfaces/defaulthorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="../interfaces/ninepatchhorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="../interfaces/defaultverticalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="../interfaces/ninepatchverticalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
 						<a href="../interfaces/scrollableparameterobject.html" class="tsd-kind-icon">Scrollable<wbr>Parameter<wbr>Object</a>
@@ -1730,7 +1730,7 @@
 						<a href="../interfaces/scrollbaroperations.html" class="tsd-kind-icon">Scrollbar<wbr>Operations</a>
 					</li>
 					<li class=" tsd-kind-type-alias tsd-is-external">
-						<a href="../globals.html#scrollbarlike" class="tsd-kind-icon">Scrollbar<wbr>Like</a>
+						<a href="../globals.html#scrollbar" class="tsd-kind-icon">Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-function tsd-is-external">
 						<a href="../globals.html#createdefaultscrollbarimage" class="tsd-kind-icon">create<wbr>Default<wbr>Scrollbar<wbr>Image</a>
