+++ docs/apiref/classes/scrolledcontentcontainer.html
@@ -180,7 +180,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides E.__constructor</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L26">Scrollable.ts:26</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L26">Scrollable.ts:26</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -202,7 +202,7 @@
 					<div class="tsd-signature tsd-kind-icon">_content<span class="tsd-signature-symbol">:</span> <a href="scrolledcontent.html" class="tsd-signature-type">ScrolledContent</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L26">Scrollable.ts:26</a></li>
+							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L26">Scrollable.ts:26</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -249,7 +249,7 @@
 					<div class="tsd-signature tsd-kind-icon">_offset<wbr>Container<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">E</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L25">Scrollable.ts:25</a></li>
+							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L25">Scrollable.ts:25</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -400,7 +400,7 @@
 					<div class="tsd-signature tsd-kind-icon">on<wbr>Content<wbr>Modified<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">Trigger</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">void</span><span class="tsd-signature-symbol">&gt;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L24">Scrollable.ts:24</a></li>
+							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L24">Scrollable.ts:24</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -875,7 +875,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L45">Scrollable.ts:45</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L45">Scrollable.ts:45</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">E</span></h4>
@@ -893,7 +893,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides E.destroy</p>
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L36">Scrollable.ts:36</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L36">Scrollable.ts:36</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">void</span></h4>
@@ -1245,7 +1245,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/Scrollable.ts#L41">Scrollable.ts:41</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollable.ts#L41">Scrollable.ts:41</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">E</span></h4>
@@ -1604,10 +1604,10 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="defaulthorizontalscrollbar.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar</a>
+						<a href="ninepatchhorizontalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="defaultverticalscrollbar.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar</a>
+						<a href="ninepatchverticalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
 						<a href="nullscrollbar.html" class="tsd-kind-icon">Null<wbr>Scrollbar</a>
@@ -1802,10 +1802,10 @@
 				</ul>
 				<ul class="after-current">
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="../interfaces/defaulthorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="../interfaces/ninepatchhorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="../interfaces/defaultverticalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="../interfaces/ninepatchverticalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
 						<a href="../interfaces/scrollableparameterobject.html" class="tsd-kind-icon">Scrollable<wbr>Parameter<wbr>Object</a>
@@ -1814,7 +1814,7 @@
 						<a href="../interfaces/scrollbaroperations.html" class="tsd-kind-icon">Scrollbar<wbr>Operations</a>
 					</li>
 					<li class=" tsd-kind-type-alias tsd-is-external">
-						<a href="../globals.html#scrollbarlike" class="tsd-kind-icon">Scrollbar<wbr>Like</a>
+						<a href="../globals.html#scrollbar" class="tsd-kind-icon">Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-function tsd-is-external">
 						<a href="../globals.html#createdefaultscrollbarimage" class="tsd-kind-icon">create<wbr>Default<wbr>Scrollbar<wbr>Image</a>
