+++ docs/apiref/index.html
@@ -67,7 +67,7 @@
 				<h1 id="akashic-scrollable">akashic-scrollable</h1>
 				<p><strong>akashic-scrollable</strong> is a library for <a href="https://akashic-games.github.io/">Akashic Engine</a> that provides scrolling/clipping entites.</p>
 				<blockquote>
-					<p>日本語の文書は [こちら][guide-ja] 。</p>
+					<p>日本語の文書は <a href="https://xnv.github.io/akashic-scrollable/book/ja/">こちら</a> 。</p>
 					<p>Unfortunately, Akashic Engine does not provide English documentation yet.
 						Whole this library is documented in English just to be prepared for their future translation.
 					You need to know Japanese to use this library in practice, at least currently.</p>
@@ -83,9 +83,9 @@
 				<h2 id="documents">Documents</h2>
 				<p>will be written later.</p>
 				<ul>
-					<li>[Guide][guide] or [利用ガイド (ja)][guide-ja]</li>
+					<li><a href="https://xnv.github.io/akashic-scrollable/book/en/">Guide</a> or <a href="https://xnv.github.io/akashic-scrollable/book/ja/">利用ガイド (ja)</a></li>
 					<li>[Demo][demo]</li>
-					<li>[API Reference][apiref]</li>
+					<li><a href="https://xnv.github.io/akashic-scrollable/apiref/index.html">API Reference</a></li>
 				</ul>
 				<h2 id="license">License</h2>
 				<p>MIT. See LICENSE for detail.</p>
@@ -102,10 +102,10 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="classes/defaulthorizontalscrollbar.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar</a>
+						<a href="classes/ninepatchhorizontalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="classes/defaultverticalscrollbar.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar</a>
+						<a href="classes/ninepatchverticalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
 						<a href="classes/nullscrollbar.html" class="tsd-kind-icon">Null<wbr>Scrollbar</a>
@@ -120,10 +120,10 @@
 						<a href="classes/scrolledcontentcontainer.html" class="tsd-kind-icon">Scrolled<wbr>Content<wbr>Container</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="interfaces/defaulthorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="interfaces/ninepatchhorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="interfaces/defaultverticalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="interfaces/ninepatchverticalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
 						<a href="interfaces/scrollableparameterobject.html" class="tsd-kind-icon">Scrollable<wbr>Parameter<wbr>Object</a>
@@ -132,7 +132,7 @@
 						<a href="interfaces/scrollbaroperations.html" class="tsd-kind-icon">Scrollbar<wbr>Operations</a>
 					</li>
 					<li class=" tsd-kind-type-alias tsd-is-external">
-						<a href="globals.html#scrollbarlike" class="tsd-kind-icon">Scrollbar<wbr>Like</a>
+						<a href="globals.html#scrollbar" class="tsd-kind-icon">Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-function tsd-is-external">
 						<a href="globals.html#createdefaultscrollbarimage" class="tsd-kind-icon">create<wbr>Default<wbr>Scrollbar<wbr>Image</a>
