+++ docs/apiref/interfaces/scrollbaroperations.html
@@ -85,8 +85,8 @@
 			<section class="tsd-panel">
 				<h3>Implemented by</h3>
 				<ul class="tsd-hierarchy">
-					<li><a href="../classes/defaulthorizontalscrollbar.html" class="tsd-signature-type">DefaultHorizontalScrollbar</a></li>
-					<li><a href="../classes/defaultverticalscrollbar.html" class="tsd-signature-type">DefaultVerticalScrollbar</a></li>
+					<li><a href="../classes/ninepatchhorizontalscrollbar.html" class="tsd-signature-type">NinePatchHorizontalScrollbar</a></li>
+					<li><a href="../classes/ninepatchverticalscrollbar.html" class="tsd-signature-type">NinePatchVerticalScrollbar</a></li>
 					<li><a href="../classes/nullscrollbar.html" class="tsd-signature-type">NullScrollbar</a></li>
 				</ul>
 			</section>
@@ -117,7 +117,7 @@
 					<div class="tsd-signature tsd-kind-icon">on<wbr>Change<wbr>Bar<wbr>Position<wbr>Rate<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">Trigger</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">&gt;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/ScrollbarLike.ts#L14">ScrollbarLike.ts:14</a></li>
+							<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollbar.ts#L14">Scrollbar.ts:14</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -142,7 +142,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/4f6e692/src/ScrollbarLike.ts#L26">ScrollbarLike.ts:26</a></li>
+									<li>Defined in <a href="https://github.com/xnv/akashic-scrollable/blob/8141b2d/src/Scrollbar.ts#L26">Scrollbar.ts:26</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -191,10 +191,10 @@
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="../classes/defaulthorizontalscrollbar.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar</a>
+						<a href="../classes/ninepatchhorizontalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
-						<a href="../classes/defaultverticalscrollbar.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar</a>
+						<a href="../classes/ninepatchverticalscrollbar.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-class tsd-is-external">
 						<a href="../classes/nullscrollbar.html" class="tsd-kind-icon">Null<wbr>Scrollbar</a>
@@ -209,10 +209,10 @@
 						<a href="../classes/scrolledcontentcontainer.html" class="tsd-kind-icon">Scrolled<wbr>Content<wbr>Container</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="defaulthorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="ninepatchhorizontalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Horizontal<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
-						<a href="defaultverticalscrollbarparameterobject.html" class="tsd-kind-icon">Default<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
+						<a href="ninepatchverticalscrollbarparameterobject.html" class="tsd-kind-icon">Nine<wbr>Patch<wbr>Vertical<wbr>Scrollbar<wbr>Parameter<wbr>Object</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-is-external">
 						<a href="scrollableparameterobject.html" class="tsd-kind-icon">Scrollable<wbr>Parameter<wbr>Object</a>
@@ -233,7 +233,7 @@
 				</ul>
 				<ul class="after-current">
 					<li class=" tsd-kind-type-alias tsd-is-external">
-						<a href="../globals.html#scrollbarlike" class="tsd-kind-icon">Scrollbar<wbr>Like</a>
+						<a href="../globals.html#scrollbar" class="tsd-kind-icon">Scrollbar</a>
 					</li>
 					<li class=" tsd-kind-function tsd-is-external">
 						<a href="../globals.html#createdefaultscrollbarimage" class="tsd-kind-icon">create<wbr>Default<wbr>Scrollbar<wbr>Image</a>
